requirements.txt
astroid==2.4.2
backcall==0.2.0
boto3==1.14.20
botocore==1.17.20
certifi==2020.6.20
chardet==3.0.4
daemon==1.2
decorator==4.4.2
docutils==0.15.2
idna==2.10
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
jmespath==0.10.0
jupyter-client==6.1.5
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.5.3
PySocks==1.7.1
python-daemon==2.2.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
tweepy==3.9.0
TwitterAPI==2.5.12
urllib3==1.25.9
utils==1.0.1
wcwidth==0.2.5
wrapt==1.12.1
